+++ pennylane_lightning/src/tests/Test_GateImplementations_Matrix.cpp
@@ -10,6 +10,7 @@
 #endif
 
 using namespace Pennylane;
+using Pennylane::Util::randomUnitary;
 
 template <typename PrecisionT>
 using ApplyMatrixType = void (*)(std::complex<PrecisionT> *, size_t,
@@ -86,7 +87,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -142,7 +143,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -198,7 +199,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -266,7 +267,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -334,7 +335,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -451,7 +452,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -760,7 +761,7 @@ void testApplyMatrix() {
         auto st = ini_st;
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, false);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 
@@ -807,7 +808,7 @@ void testApplyMatrixInverse() {
                                         wires, false);
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -825,7 +826,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -843,7 +844,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -861,7 +862,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -879,7 +880,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", wires = {1,2} - "
@@ -895,7 +896,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", wires = {1,3} - "
@@ -911,7 +912,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -928,7 +929,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", wires = {0,1,2,3} - "
@@ -944,7 +945,7 @@ void testApplyMatrixInverse() {
         GateImplementation::applyMatrix(st.data(), num_qubits, matrix.data(),
                                         wires, true);
 
-        REQUIRE(st == PLApprox(ini_st).margin(margin));
+        REQUIRE(st == approx(ini_st).margin(margin));
     }
 }