+++ pennylane_lightning/src/tests/Test_GateImplementations_Nonparam.cpp
@@ -21,10 +21,6 @@
  */
 using namespace Pennylane;
 
-namespace {
-using std::vector;
-}
-
 /**
  * @brief Run test suit only when the gate is defined
  */
@@ -104,7 +100,7 @@ void testApplyPauliY() {
 
         GateImplementation::applyPauliY(st.data(), num_qubits, {index}, false);
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(PauliY);
@@ -127,7 +123,7 @@ void testApplyPauliZ() {
         auto st = createPlusState<PrecisionT>(num_qubits);
         GateImplementation::applyPauliZ(st.data(), num_qubits, {index}, false);
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(PauliZ);
@@ -173,7 +169,7 @@ template <typename PrecisionT, class GateImplementation> void testApplyS() {
 
         GateImplementation::applyS(st.data(), num_qubits, {index}, false);
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(S);
@@ -196,7 +192,7 @@ template <typename PrecisionT, class GateImplementation> void testApplyT() {
 
         GateImplementation::applyT(st.data(), num_qubits, {index}, false);
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(T);