+++ pennylane_lightning/src/tests/Test_GateImplementations_Param.cpp
@@ -111,7 +111,7 @@ void testApplyPhaseShift() {
         GateImplementation::applyPhaseShift(st.data(), num_qubits, {index},
                                             false, {angles[index]});
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(PhaseShift);
@@ -136,7 +136,7 @@ void testApplyRX() {
         GateImplementation::applyRX(st.data(), num_qubits, {0}, false,
                                     {angles[index]});
 
-        CHECK(st == PLApprox(expected_results[index]).epsilon(1e-7));
+        CHECK(st == approx(expected_results[index]).epsilon(1e-7));
     }
 }
 PENNYLANE_RUN_TEST(RX);
@@ -176,7 +176,7 @@ void testApplyRY() {
             auto st = init_state;
             GateImplementation::applyRY(st.data(), num_qubits, {0}, false,
                                         {angles[index]});
-            CHECK(st == PLApprox(expected_results[index]).epsilon(1e-5));
+            CHECK(st == approx(expected_results[index]).epsilon(1e-5));
         }
     }
 }
@@ -225,7 +225,7 @@ void testApplyRZ() {
         GateImplementation::applyRZ(st.data(), num_qubits, {index}, false,
                                     {angles[index]});
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(RZ);
@@ -242,9 +242,9 @@ void testApplyRot() {
         std::vector<PrecisionT>{2.3, 0.1, 0.4}};
 
     std::vector<std::vector<ComplexPrecisionT>> expected_results{
-        std::vector<ComplexPrecisionT>(0b1 << num_qubits),
-        std::vector<ComplexPrecisionT>(0b1 << num_qubits),
-        std::vector<ComplexPrecisionT>(0b1 << num_qubits)};
+        std::vector<ComplexPrecisionT>(1U << num_qubits),
+        std::vector<ComplexPrecisionT>(1U << num_qubits),
+        std::vector<ComplexPrecisionT>(1U << num_qubits)};
 
     for (size_t i = 0; i < angles.size(); i++) {
         const auto rot_mat =
@@ -259,7 +259,7 @@ void testApplyRot() {
                                      angles[index][0], angles[index][1],
                                      angles[index][2]);
 
-        CHECK(st == PLApprox(expected_results[index]));
+        CHECK(st == approx(expected_results[index]));
     }
 }
 PENNYLANE_RUN_TEST(Rot);
@@ -294,7 +294,7 @@ void testApplyIsingXX() {
         auto st = ini_st;
         GateImplementation::applyIsingXX(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingXX0,1 |100> -> a|100> + b|010> - "
@@ -317,7 +317,7 @@ void testApplyIsingXX() {
         auto st = ini_st;
         GateImplementation::applyIsingXX(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingXX0,1 |010> -> a|010> + b|100> - "
@@ -340,7 +340,7 @@ void testApplyIsingXX() {
         auto st = ini_st;
         GateImplementation::applyIsingXX(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingXX0,1 |110> -> a|110> + b|000> - "
@@ -363,7 +363,7 @@ void testApplyIsingXX() {
         auto st = ini_st;
         GateImplementation::applyIsingXX(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingXX0,2 - "
@@ -395,7 +395,7 @@ void testApplyIsingXX() {
         auto st = ini_st;
         GateImplementation::applyIsingXX(st.data(), num_qubits, wires, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(IsingXX);
@@ -427,7 +427,7 @@ void testApplyIsingYY() {
         auto st = ini_st;
         GateImplementation::applyIsingYY(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingYY0,1 |100> -> a|100> + b|010> - "
@@ -450,7 +450,7 @@ void testApplyIsingYY() {
         auto st = ini_st;
         GateImplementation::applyIsingYY(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingYY0,1 |010> -> a|010> + b|100> - "
@@ -473,7 +473,7 @@ void testApplyIsingYY() {
         auto st = ini_st;
         GateImplementation::applyIsingYY(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingYY0,1 |110> -> a|110> + b|000> - "
@@ -496,7 +496,7 @@ void testApplyIsingYY() {
         auto st = ini_st;
         GateImplementation::applyIsingYY(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingYY0,1 - "
@@ -547,7 +547,7 @@ void testApplyIsingYY() {
         auto st = ini_st;
         GateImplementation::applyIsingYY(st.data(), num_qubits, wires, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(IsingYY);
@@ -579,7 +579,7 @@ void testApplyIsingZZ() {
         auto st = ini_st;
         GateImplementation::applyIsingZZ(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingZZ0,1 |100> -> |100> - "
@@ -602,7 +602,7 @@ void testApplyIsingZZ() {
         auto st = ini_st;
         GateImplementation::applyIsingZZ(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -626,7 +626,7 @@ void testApplyIsingZZ() {
         auto st = ini_st;
         GateImplementation::applyIsingZZ(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -650,7 +650,7 @@ void testApplyIsingZZ() {
         auto st = ini_st;
         GateImplementation::applyIsingZZ(st.data(), num_qubits, {0, 1}, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected_results).margin(1e-7));
+        REQUIRE(st == approx(expected_results).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", IsingZZ0,1 - "
@@ -701,7 +701,7 @@ void testApplyIsingZZ() {
         auto st = ini_st;
         GateImplementation::applyIsingZZ(st.data(), num_qubits, wires, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(IsingZZ);
@@ -741,7 +741,7 @@ void testApplyControlledPhaseShift() {
     GateImplementation::applyControlledPhaseShift(st.data(), num_qubits, {0, 1},
                                                   false, angles[0]);
     CAPTURE(st);
-    CHECK(st == PLApprox(expected_results[0]));
+    CHECK(st == approx(expected_results[0]));
 }
 PENNYLANE_RUN_TEST(ControlledPhaseShift);
 
@@ -796,7 +796,7 @@ void testApplyCRX() {
         auto st = ini_st;
         GateImplementation::applyCRX(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", CRX0,2 - " << PrecisionToName<PrecisionT>::value) {
@@ -846,7 +846,7 @@ void testApplyCRX() {
         auto st = ini_st;
         GateImplementation::applyCRX(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", CRX1,3 - " << PrecisionToName<PrecisionT>::value) {
@@ -896,7 +896,7 @@ void testApplyCRX() {
         auto st = ini_st;
         GateImplementation::applyCRX(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(CRX);
@@ -953,7 +953,7 @@ void testApplyCRY() {
         auto st = ini_st;
         GateImplementation::applyCRY(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1004,7 +1004,7 @@ void testApplyCRY() {
         auto st = ini_st;
         GateImplementation::applyCRY(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1055,7 +1055,7 @@ void testApplyCRY() {
         auto st = ini_st;
         GateImplementation::applyCRY(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 
@@ -1113,7 +1113,7 @@ void testApplyCRZ() {
         auto st = ini_st;
         GateImplementation::applyCRZ(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1164,7 +1164,7 @@ void testApplyCRZ() {
         auto st = ini_st;
         GateImplementation::applyCRZ(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1215,7 +1215,7 @@ void testApplyCRZ() {
         auto st = ini_st;
         GateImplementation::applyCRZ(st.data(), num_qubits, wires, false,
                                      angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(CRZ);
@@ -1236,7 +1236,7 @@ void testApplyCRot() {
         GateImplementation::applyCRot(st.data(), num_qubits, {0, 1}, false,
                                       angles[0], angles[1], angles[2]);
 
-        CHECK(st == PLApprox(ini_st));
+        CHECK(st == approx(ini_st));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", CRot0,1 |100> -> |1>(a|0>+b|1>)|0> - "
@@ -1256,7 +1256,7 @@ void testApplyCRot() {
         GateImplementation::applyCRot(st.data(), num_qubits, {0, 1}, false,
                                       angles[0], angles[1], angles[2]);
 
-        CHECK(st == PLApprox(expected_results));
+        CHECK(st == approx(expected_results));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1309,7 +1309,7 @@ void testApplyCRot() {
         auto st = ini_st;
         GateImplementation::applyCRot(st.data(), num_qubits, wires, false, phi,
                                       theta, omega);
-        REQUIRE(st == PLApprox(expected).margin(1e-5));
+        REQUIRE(st == approx(expected).margin(1e-5));
     }
 }
 PENNYLANE_RUN_TEST(CRot);
@@ -1342,7 +1342,7 @@ void testApplyMultiRZ() {
         GateImplementation::applyMultiRZ(st.data(), num_qubits, {0}, false,
                                          angle);
 
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", MultiRZ0 |++++> - "
@@ -1365,7 +1365,7 @@ void testApplyMultiRZ() {
         GateImplementation::applyMultiRZ(st.data(), num_qubits, {0}, false,
                                          angle);
 
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", MultiRZ01 |++++> - "
@@ -1388,7 +1388,7 @@ void testApplyMultiRZ() {
         GateImplementation::applyMultiRZ(st.data(), num_qubits, {0, 1}, false,
                                          angle);
 
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", MultiRZ012 |++++> - "
@@ -1411,7 +1411,7 @@ void testApplyMultiRZ() {
         GateImplementation::applyMultiRZ(st.data(), num_qubits, {0, 1, 2},
                                          false, angle);
 
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
     DYNAMIC_SECTION(GateImplementation::name
                     << ", MultiRZ0123 |++++> - "
@@ -1434,7 +1434,7 @@ void testApplyMultiRZ() {
         GateImplementation::applyMultiRZ(st.data(), num_qubits, {0, 1, 2, 3},
                                          false, angle);
 
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
 
     DYNAMIC_SECTION(GateImplementation::name
@@ -1484,7 +1484,7 @@ void testApplyMultiRZ() {
 
         GateImplementation::applyMultiRZ(st.data(), num_qubits, wires, false,
                                          angle);
-        REQUIRE(st == PLApprox(expected).margin(1e-7));
+        REQUIRE(st == approx(expected).margin(1e-7));
     }
 }
 PENNYLANE_RUN_TEST(MultiRZ);