+++ pennylane_lightning/src/tests/Test_Util.cpp
@@ -64,6 +64,7 @@ TEMPLATE_TEST_CASE("Constant values", "[Util]", float, double) {
 // NOLINTNEXTLINE: Avoid complexity errors
 TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                    double) {
+    using Util::Trans;
     SECTION("exp2: 2^n") {
         for (size_t i = 0; i < 10; i++) {
             CHECK(Util::exp2(i) == static_cast<size_t>(std::pow(2, i)));
@@ -168,7 +169,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 CAPTURE(v_out);
                 CAPTURE(v_expected);
 
-                CHECK(v_out == PLApprox(v_expected).margin(1e-7));
+                CHECK(v_out == approx(v_expected).margin(1e-7));
             }
         }
         SECTION("Random Complex") {
@@ -194,7 +195,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 Util::matrixVecProd(mat, v_in, 4, 4);
             CAPTURE(v_out);
 
-            CHECK(v_out == PLApprox(v_expected).margin(1e-7));
+            CHECK(v_out == approx(v_expected).margin(1e-7));
         }
         SECTION("Invalid Arguments") {
             using namespace Catch::Matchers;
@@ -224,7 +225,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 CAPTURE(v_out);
                 CAPTURE(v_expected);
 
-                CHECK(v_out == PLApprox(v_expected).margin(1e-7));
+                CHECK(v_out == approx(v_expected).margin(1e-7));
             }
         }
         SECTION("Zero Vector") {
@@ -238,7 +239,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 CAPTURE(v_out);
                 CAPTURE(v_expected);
 
-                CHECK(v_out == PLApprox(v_expected).margin(1e-7));
+                CHECK(v_out == approx(v_expected).margin(1e-7));
             }
         }
         SECTION("Random Matrix") {
@@ -251,7 +252,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
             CAPTURE(v_out);
             CAPTURE(v_expected);
 
-            CHECK(v_out == PLApprox(v_expected).margin(1e-7));
+            CHECK(v_out == approx(v_expected).margin(1e-7));
         }
     }
     SECTION("Transpose") {
@@ -267,7 +268,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 CAPTURE(mat_t);
                 CAPTURE(mat);
 
-                CHECK(mat_t == PLApprox(mat).margin(1e-7));
+                CHECK(mat_t == approx(mat).margin(1e-7));
             }
         }
         SECTION("Random Complex") {
@@ -295,7 +296,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
             CAPTURE(mat_t);
             CAPTURE(mat_t_exp);
 
-            CHECK(mat_t == PLApprox(mat_t_exp));
+            CHECK(mat_t == approx(mat_t_exp));
         }
         SECTION("Invalid Arguments") {
             using namespace Catch::Matchers;
@@ -320,7 +321,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                 CAPTURE(m_out);
                 CAPTURE(m_out_exp);
 
-                CHECK(m_out == PLApprox(m_out_exp));
+                CHECK(m_out == approx(m_out_exp));
             }
         }
         SECTION("Random Complex") {
@@ -377,8 +378,8 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
             CAPTURE(m_out_2);
             CAPTURE(m_out_exp);
 
-            CHECK(m_out_1 == PLApprox(m_out_2));
-            CHECK(m_out_1 == PLApprox(m_out_exp));
+            CHECK(m_out_1 == approx(m_out_2));
+            CHECK(m_out_1 == approx(m_out_exp));
         }
         SECTION("Random complex non-square") {
             const size_t m = 4;
@@ -451,7 +452,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
 
             const auto m_out = Util::matrixMatProd(mat1, mat2, m, n, k);
 
-            CHECK(m_out == PLApprox(expected));
+            CHECK(m_out == approx(expected));
         }
         SECTION("Invalid Arguments") {
             using namespace Catch::Matchers;
@@ -469,6 +470,17 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
                                        "the input right matrix"));
         }
     }
+    SECTION("SquaredNorm") {
+        SECTION("For real type") {
+            std::vector<TestType> vec{0.0, 1.0, 3.0, 10.0};
+            CHECK(Util::squaredNorm(vec) == Approx(110.0));
+        }
+
+        SECTION("For complex type") {
+            std::vector<std::complex<TestType>> vec{{0.0, 1.0}, {3.0, 10.0}};
+            CHECK(Util::squaredNorm(vec) == Approx(110.0));
+        }
+    }
 }
 
 /**
@@ -478,7 +490,7 @@ TEMPLATE_TEST_CASE("Utility math functions", "[Util][LinearAlgebra]", float,
  */
 size_t popcount_slow(uint64_t x) {
     size_t c = 0;
-    for (; x != 0; x >>= 1) {
+    for (; x != 0U; x >>= 1U) {
         if ((x & 1U) != 0U) {
             c++;
         }
@@ -493,8 +505,8 @@ size_t popcount_slow(uint64_t x) {
  */
 size_t ctz_slow(uint64_t x) {
     size_t c = 0;
-    while ((x & 1) == 0) {
-        x >>= 1;
+    while ((x & 1U) == 0) {
+        x >>= 1U;
         c++;
     }
     return c;
@@ -603,4 +615,36 @@ TEST_CASE("Utility array and tuples", "[Util]") {
                 std::pair<std::string_view, int>("Three", 3),
                 std::pair<std::string_view, int>("Four", 4),
             });
+}
+
+/**
+ * @brief Test randomUnitary is correct
+ */
+TEMPLATE_TEST_CASE("randomUnitary", "[Test_Internal]", float, double) {
+    using PrecisionT = TestType;
+
+    std::mt19937 re{1337};
+
+    for (size_t num_qubits = 1; num_qubits <= 5; num_qubits++) {
+        const size_t dim = (1U << num_qubits);
+        const auto unitary = Util::randomUnitary<PrecisionT>(re, num_qubits);
+
+        auto unitary_dagger = Util::Transpose(unitary, dim, dim);
+        std::transform(
+            unitary_dagger.begin(), unitary_dagger.end(),
+            unitary_dagger.begin(),
+            [](const std::complex<PrecisionT> &v) { return std::conj(v); });
+
+        std::vector<std::complex<PrecisionT>> mat(dim * dim);
+        Util::matrixMatProd(unitary.data(), unitary_dagger.data(), mat.data(),
+                            dim, dim, dim);
+
+        std::vector<std::complex<PrecisionT>> identity(
+            dim * dim, std::complex<PrecisionT>{});
+        for (size_t i = 0; i < dim; i++) {
+            identity[i * dim + i] = std::complex<PrecisionT>{1.0, 0.0};
+        }
+
+        REQUIRE(mat == approx(identity).margin(1e-5));
+    }
 }